parting commit
this commit contains the final state of the game before i turned it in. technically, i'm committing this on sunday two days after the deadline but the game is unchanged since then. i think i'm gonna take a little break from development for a while but i'm excited to keep working on this.